refactor(notes): extract types and utilities from NotesManagerPage monolith
Move ~1040 lines of type definitions and pure utility functions from
NotesManagerPage.tsx into dedicated modules:
- notes-manager-types.ts: re-exports all type definitions
- notes-manager-utils.ts: all pure utility functions and type definitions

This is the first step of the Notes page component decomposition.
NotesManagerPage.tsx reduced from 8801 to 7852 lines. All 137 tests pass.
